Added auth

+++ src/main/java/com/oop/labs/manual/dto/Function.java
@@ -56,7 +56,7 @@ public class Function {
         return authorId;
     }
 
-    public void setAuthorId(int authorId) {
+    public void setAuthorId(long authorId) {
         this.authorId = authorId;
     }
